[Refactor] Reorganised code structure
