+++ bertrend/demos/topic_analysis/data_utils.py
@@ -10,9 +10,9 @@ import pandas as pd
 import streamlit as st
 from pathlib import Path
 
-from bertrend.topic_analysis.app_utils import plot_docs_reparition_over_time
-from bertrend.topic_analysis.state_utils import save_widget_state
-from bertrend.utils import TEXT_COLUMN, TIMESTAMP_COLUMN
+from bertrend.demos.topic_analysis.app_utils import plot_docs_reparition_over_time
+from bertrend.demos.topic_analysis.state_utils import save_widget_state
+from bertrend.utils.data_loading import TEXT_COLUMN, TIMESTAMP_COLUMN
 
 
 def data_overview(df: pd.DataFrame):
