+++ bertrend/demos/topic_analysis/pages/4_Generate_Newsletters.py
@@ -7,7 +7,7 @@ import inspect
 import streamlit as st
 from pathlib import Path
 
-from bertrend.topic_analysis.state_utils import (
+from bertrend.demos.topic_analysis.state_utils import (
     restore_widget_state,
     register_widget,
     save_widget_state,
