+++ bertrend/demos/topic_analysis/pages/hidden/5_Merge_models.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from bertopic import BERTopic
 from typing import List, Optional, Union
 
-from bertrend.topic_analysis.state_utils import restore_widget_state
+from bertrend.demos.topic_analysis.state_utils import restore_widget_state
 
 
 def list_saved_models(saved_models_dir: Union[str, Path]) -> List[Path]:
