+++ bertrend/demos/topic_analysis/pages/hidden/7_Temporal_Visualizations.py
@@ -9,20 +9,20 @@ import locale
 # from langchain_openai import ChatOpenAI
 # from langchain_core.messages import HumanMessage, SystemMessage
 
-from bertrend.topic_analysis.app_utils import (
+from bertrend.demos.topic_analysis.app_utils import (
     plot_topics_over_time,
     compute_topics_over_time,
 )
 
 import pandas as pd
 
-from bertrend.topic_analysis.state_utils import (
+from bertrend.demos.topic_analysis.state_utils import (
     restore_widget_state,
     register_widget,
     save_widget_state,
 )
 from bertrend.metrics.temporal_metrics import TempTopic
-from bertrend.utils import TEXT_COLUMN, TIMESTAMP_COLUMN
+from bertrend.utils.data_loading import TIMESTAMP_COLUMN, TEXT_COLUMN
 
 
 def display_documents_on_click(clicked_point):
