+++ bertrend/metrics/temporal_metrics_embedding.py
@@ -983,7 +983,7 @@ class TempTopic:
         - Replacing hyphens and similar characters with spaces
         - Removing specific prefixes
         - Removing all punctuation
-        - Replacing special characters with spaces (preserving accented characters, common Latin extensions, and newlines)
+        - Replacing special characters with spaces (preserving accented characters, llm_utils Latin extensions, and newlines)
         - Normalizing superscripts and subscripts
         - Splitting words containing capitals in the middle (while avoiding splitting fully capitalized words)
         - Lowercasing all text
