+++ bertrend/services/embedding_service.py
