+++ bertrend/summary/chatgpt_summarizer.py
@@ -5,7 +5,7 @@
 
 from loguru import logger
 
-from bertrend.common.openai_client import OpenAI_Client
+from bertrend.llm_utils.openai_client import OpenAI_Client
 from bertrend.summary.prompts import (
     FR_SYSTEM_SUMMARY_SENTENCES,
     EN_SYSTEM_SUMMARY_SENTENCES,
