+++ bertrend/trend_analysis/__init__.py
