+++ bertrend/trend_analysis/signal_llm_template_en.html
