+++ bertrend/trend_analysis/signal_llm_template_fr.html
