+++ bertrend/utils/__init__.py
@@ -2,3 +2,19 @@
 #  See AUTHORS.txt
 #  SPDX-License-Identifier: MPL-2.0
 #  This file is part of BERTrend.
+
+import os
+import ssl
+import nltk
+
+# Ensures files are written with +rw permissions for both user and groups
+os.umask(0o002)
+
+# Workaround for downloading nltk data in some environments
+try:
+    _create_unverified_https_context = ssl._create_unverified_context
+except AttributeError:
+    pass
+else:
+    ssl._create_default_https_context = _create_unverified_https_context
+nltk.download("stopwords")
