+++ bertrend_apps/newsletters/__main__.py
@@ -25,17 +25,17 @@ from umap import UMAP
 
 from bertrend import FEED_BASE_PATH, BEST_CUDA_DEVICE, OUTPUT_PATH
 from bertrend.utils.config_utils import parse_literal, EnvInterpolation
+from bertrend.utils.data_loading import (
+    split_df_by_paragraphs,
+    TIMESTAMP_COLUMN,
+    TEXT_COLUMN,
+    load_data,
+)
 from bertrend_apps.newsletters.newsletter_features import (
     generate_newsletter,
     export_md_string,
 )
 from bertrend.train import EmbeddingModel, train_BERTopic
-from bertrend.utils import (
-    load_data,
-    TEXT_COLUMN,
-    TIMESTAMP_COLUMN,
-    split_df_by_paragraphs,
-)
 from bertrend_apps.common.mail_utils import get_credentials, send_email
 from bertrend_apps.common.crontab_utils import schedule_newsletter
