+++ pyproject.toml
@@ -17,25 +17,13 @@ exclude = [ "data" ]
 python = "^3.11"
 ipython = "^8.28.0"
 
-# A list of all the optional dependencies, some of which are included in the
-# below `extras`. They can be opted into by apps.
-pytest = { version = "^8.3.3", optional = true}
-coverage = { version = "^7.6.9", optional = true}
-
 # These packages are mandatory and form the core of this package’s distribution.
 accelerate = "^0.34.2"
-arxiv = "^2.1.3"
 bertopic = "0.16.2" #problem with versions above in .update_topics
 black = "^24.10.0"
 dateparser = "^1.2.0"
 datamapplot = "0.3.0"
-feedparser = "^6.0.11"
 gensim = "4.3.2"
-googlenewsdecoder = "^0.1.6"
-google-api-python-client = "^2.148.0"
-google-auth-httplib2 = "^0.2.0"
-google-auth-oauthlib = "^1.2.1"
-goose3 = "^3.1.19"
 hdbscan = "^0.8.40"
 langchain-community = "^0.3.1"
 langchain-chroma = "^0.1.4"
@@ -48,15 +36,12 @@ loguru = "^0.7.3"
 lxml-html-clean = "^0.2.2"
 jsonlines = "^4.0.0"
 markdown = "^3.7"
-newscatcherapi = "^0.7.3"
-newspaper4k = "^0.9.3.1"
 nltk = "^3.9.1"
 numpy = "<2"
 openai = "^1.57.2"
 pandas = "^2.2.2"
 plotly = "^5.24.1"
 plotly-resampler = "^0.10.0"
-pygooglenews = {git = "https://github.com/picaultj/pygooglenews.git", rev = "master"}
 scipy = {version = "1.12", python=""}
 scikit-learn = "^1.5.2"
 sentence-transformers = "3.0.1"
@@ -71,9 +56,28 @@ tiktoken = "^0.8.0"
 tokenizers = "<0.20.0"
 umap-learn = "^0.5.7"
 
+# A list of all the optional dependencies, some of which are included in the
+# below `extras`. They can be opted into by apps.
+# tests
+pytest = { version = "^8.3.3", optional = true}
+coverage = { version = "^7.6.9", optional = true}
+# BERTrend apps
+arxiv = { version = "^2.1.3", optional = true }
+feedparser = { version = "^6.0.11", optional = true }
+googlenewsdecoder = { version = "^0.1.6", optional = true }
+google-api-python-client = { version = "^2.148.0", optional = true }
+google-auth-httplib2 = { version = "^0.2.0", optional = true }
+google-auth-oauthlib = { version = "^1.2.1", optional = true }
+goose3 = { version = "^3.1.19", optional = true }
+newscatcherapi = { version = "^0.7.3", optional = true }
+newspaper4k = { version = "^0.9.3.1", optional = true }
+pygooglenews = {git = "https://github.com/picaultj/pygooglenews.git", rev = "master", optional = true}
+
 
 [tool.poetry.extras]
 tests = ["pytest", "coverage"]
+apps = ["arxiv", "feedparser", "googlenewsdecoder", "google-api-python-client", "google-auth-httplib2",
+        "google-auth-oauthlib", "goose3", "newscatcherapi", "newspaper4k", "pygooglenews"]
 
 
 [build-system]
